+++ cmd/ecom-api/main.go
@@ -37,7 +37,7 @@ import (
 	"google.golang.org/api/option"
 )
 
-var version = "v0.60.0"
+var version = "v0.60.1"
 
 const maxDbConnectAttempts = 3