+++ schemas/postgres/schema/shipping_tariff.sql
@@ -2,7 +2,7 @@ CREATE TABLE shipping_tariff (
   id             SERIAL PRIMARY KEY,
   uuid           UUID UNIQUE NOT NULL DEFAULT uuid_generate_v4(),
   country_code   CHAR(2) NOT NULL,
-  shipping_code  VARCHAR(256) NOT NULL UNIQUE,
+  shipping_code  VARCHAR(256) NOT NULL,
   name           VARCHAR(512),
   price          INTEGER NOT NULL CHECK (price >= 0),
   tax_code       VARCHAR(32) NULL DEFAULT NULL,